requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
requests>=2.0.0
beautifulsoup4>=4.0.0
python-dateutil>=2.0.0
lxml>=4.0.0